+++ AdMob/MPGoogleAdMobNativeCustomEvent.m
@@ -1,12 +1,15 @@
-#import "MPGoogleAdMobNativeAdAdapter.h"
 #import "MPGoogleAdMobNativeCustomEvent.h"
+#import "MPGoogleAdMobNativeAdAdapter.h"
 #if __has_include("MoPub.h")
-    #import "MPLogging.h"
-    #import "MPNativeAd.h"
-    #import "MPNativeAdConstants.h"
-    #import "MPNativeAdError.h"
-    #import "MPNativeAdUtils.h"
+#import "MPLogging.h"
+#import "MPNativeAd.h"
+#import "MPNativeAdConstants.h"
+#import "MPNativeAdError.h"
+#import "MPNativeAdUtils.h"
 #endif
+
+
+#import "UIView+MPGoogleAdMobAdditions.h"
 
 static void MPGoogleLogInfo(NSString *message) {
   message = [[NSString alloc] initWithFormat:@"<Google Adapter> - %@", message];
@@ -16,8 +19,7 @@ static void MPGoogleLogInfo(NSString *message) {
 /// Holds the preferred location of the AdChoices icon.
 static GADAdChoicesPosition adChoicesPosition;
 
-@interface MPGoogleAdMobNativeCustomEvent () <
-    GADAdLoaderDelegate, GADNativeAppInstallAdLoaderDelegate, GADNativeContentAdLoaderDelegate>
+@interface MPGoogleAdMobNativeCustomEvent () <GADAdLoaderDelegate, GADUnifiedNativeAdLoaderDelegate>
 
 /// GADAdLoader instance.
 @property(nonatomic, strong) GADAdLoader *adLoader;
@@ -60,7 +62,6 @@ static GADAdChoicesPosition adChoicesPosition;
   request.requestAgent = @"MoPub";
   GADNativeAdImageAdLoaderOptions *nativeAdImageLoaderOptions =
       [[GADNativeAdImageAdLoaderOptions alloc] init];
-  nativeAdImageLoaderOptions.disableImageLoading = YES;
   nativeAdImageLoaderOptions.shouldRequestMultipleImages = NO;
   nativeAdImageLoaderOptions.preferredImageOrientation =
       GADNativeAdImageAdLoaderOptionsOrientationAny;
@@ -70,23 +71,27 @@ static GADAdChoicesPosition adChoicesPosition;
   GADNativeAdViewAdOptions *nativeAdViewAdOptions = [[GADNativeAdViewAdOptions alloc] init];
   nativeAdViewAdOptions.preferredAdChoicesPosition = adChoicesPosition;
 
-  self.adLoader = [[GADAdLoader alloc]
-        initWithAdUnitID:adUnitID
-      rootViewController:rootViewController
-                 adTypes:@[ kGADAdLoaderAdTypeNativeAppInstall, kGADAdLoaderAdTypeNativeContent ]
-                 options:@[ nativeAdImageLoaderOptions, nativeAdViewAdOptions ]];
+  self.adLoader =
+      [[GADAdLoader alloc] initWithAdUnitID:adUnitID
+                         rootViewController:rootViewController
+                                    adTypes:@[ kGADAdLoaderAdTypeUnifiedNative ]
+                                    options:@[ nativeAdImageLoaderOptions, nativeAdViewAdOptions ]];
   self.adLoader.delegate = self;
-    
-  // Consent collected from the MoPub’s consent dialogue should not be used to set up Google's personalization preference. Publishers should work with Google to be GDPR-compliant.
-    
-  MPGoogleGlobalMediationSettings *medSettings = [[MoPub sharedInstance] globalMediationSettingsForClass:[MPGoogleGlobalMediationSettings class]];
-    
+
+
+  // Consent collected from the MoPub’s consent dialogue should not be used to set up
+  // Google's personalization preference. Publishers should work with Google to be GDPR-compliant.
+
+
+  MPGoogleGlobalMediationSettings *medSettings = [[MoPub sharedInstance]
+      globalMediationSettingsForClass:[MPGoogleGlobalMediationSettings class]];
+
   if (medSettings.npa) {
-      GADExtras *extras = [[GADExtras alloc] init];
-      extras.additionalParameters = @{@"npa": medSettings.npa};
-      [request registerAdNetworkExtras:extras];
+    GADExtras *extras = [[GADExtras alloc] init];
+    extras.additionalParameters = @{@"npa" : medSettings.npa};
+    [request registerAdNetworkExtras:extras];
   }
-    
+
   [self.adLoader loadRequest:request];
 }
 
@@ -96,34 +101,68 @@ static GADAdChoicesPosition adChoicesPosition;
   [self.delegate nativeCustomEvent:self didFailToLoadAdWithError:error];
 }
 
-#pragma mark GADNativeAppInstallAdLoaderDelegate implementation
+#pragma mark GADUnifiedNativeAdLoaderDelegate implementation
 
-- (void)adLoader:(GADAdLoader *)adLoader
-    didReceiveNativeAppInstallAd:(GADNativeAppInstallAd *)nativeAppInstallAd {
-  if (![self isValidAppInstallAd:nativeAppInstallAd]) {
-    MPGoogleLogInfo(@"App install ad is missing one or more required assets, failing the request");
+- (void)adLoader:(nonnull GADAdLoader *)adLoader
+    didReceiveUnifiedNativeAd:(nonnull GADUnifiedNativeAd *)nativeAd {
+  if (![self isValidUnifiedNativeAd:nativeAd]) {
+    MPGoogleLogInfo(
+        @"Unified native ad is missing one or more required assets, failing the request");
     [self.delegate nativeCustomEvent:self
             didFailToLoadAdWithError:MPNativeAdNSErrorForInvalidAdServerResponse(
                                          @"Missing one or more required assets.")];
     return;
   }
 
+  GADUnifiedNativeAdView *gadUnifiedNativeAdView = [[GADUnifiedNativeAdView alloc] init];
+
+  GADAdChoicesView *adChoicesView = [[GADAdChoicesView alloc] initWithFrame:CGRectZero];
+  [gadUnifiedNativeAdView addSubview:adChoicesView];
+  gadUnifiedNativeAdView.adChoicesView = adChoicesView;
+
+  GADMediaView *mediaView = [[GADMediaView alloc] initWithFrame:CGRectZero];
+  [gadUnifiedNativeAdView addSubview:mediaView];
+  gadUnifiedNativeAdView.mediaView = mediaView;
+
+  gadUnifiedNativeAdView.nativeAd = nativeAd;
+
+  UILabel *headlineView = [[UILabel alloc] initWithFrame:CGRectZero];
+  headlineView.text = nativeAd.headline;
+  headlineView.textColor = [UIColor clearColor];
+  [gadUnifiedNativeAdView addSubview:headlineView];
+  gadUnifiedNativeAdView.headlineView = headlineView;
+
+  UILabel *bodyView = [[UILabel alloc] initWithFrame:CGRectZero];
+  bodyView.text = nativeAd.body;
+  bodyView.textColor = [UIColor clearColor];
+  [gadUnifiedNativeAdView addSubview:bodyView];
+  gadUnifiedNativeAdView.bodyView = bodyView;
+
+  UILabel *callToActionView = [[UILabel alloc] initWithFrame:CGRectZero];
+  callToActionView.text = nativeAd.callToAction;
+  callToActionView.textColor = [UIColor clearColor];
+  [gadUnifiedNativeAdView addSubview:callToActionView];
+  gadUnifiedNativeAdView.callToActionView = callToActionView;
+
+  UIImageView *mainMediaImageView = [[UIImageView alloc] initWithFrame:CGRectZero];
+  mainMediaImageView.image = nativeAd.images.firstObject.image;
+  [gadUnifiedNativeAdView addSubview:mainMediaImageView];
+  gadUnifiedNativeAdView.imageView = mainMediaImageView;
+
+  UIImageView *iconView = [[UIImageView alloc] initWithFrame:CGRectZero];
+  iconView.image = nativeAd.icon.image;
+  [gadUnifiedNativeAdView addSubview:iconView];
+  gadUnifiedNativeAdView.iconView = iconView;
+
   MPGoogleAdMobNativeAdAdapter *adapter =
-      [[MPGoogleAdMobNativeAdAdapter alloc] initWithAdMobNativeAppInstallAd:nativeAppInstallAd];
+      [[MPGoogleAdMobNativeAdAdapter alloc] initWithAdMobUnifiedNativeAd:nativeAd
+                                                     unifiedNativeAdView:gadUnifiedNativeAdView];
   MPNativeAd *moPubNativeAd = [[MPNativeAd alloc] initWithAdAdapter:adapter];
 
   NSMutableArray *imageURLs = [NSMutableArray array];
 
   if ([moPubNativeAd.properties[kAdIconImageKey] length]) {
     if (![MPNativeAdUtils addURLString:moPubNativeAd.properties[kAdIconImageKey]
-                            toURLArray:imageURLs]) {
-      [self.delegate nativeCustomEvent:self
-              didFailToLoadAdWithError:MPNativeAdNSErrorForInvalidImageURL()];
-    }
-  }
-
-  if ([moPubNativeAd.properties[kAdMainImageKey] length]) {
-    if (![MPNativeAdUtils addURLString:moPubNativeAd.properties[kAdMainImageKey]
                             toURLArray:imageURLs]) {
       [self.delegate nativeCustomEvent:self
               didFailToLoadAdWithError:MPNativeAdNSErrorForInvalidImageURL()];
@@ -141,62 +180,13 @@ static GADAdChoicesPosition adChoicesPosition;
                 }];
 }
 
-#pragma mark GADNativeContentAdLoaderDelegate implementation
-
-- (void)adLoader:(GADAdLoader *)adLoader
-    didReceiveNativeContentAd:(GADNativeContentAd *)nativeContentAd {
-  if (![self isValidContentAd:nativeContentAd]) {
-    MPGoogleLogInfo(@"Content ad is missing one or more required assets, failing the request");
-    [self.delegate nativeCustomEvent:self
-            didFailToLoadAdWithError:MPNativeAdNSErrorForInvalidAdServerResponse(
-                                         @"Missing one or more required assets.")];
-    return;
-  }
-
-  MPGoogleAdMobNativeAdAdapter *adapter =
-      [[MPGoogleAdMobNativeAdAdapter alloc] initWithAdMobNativeContentAd:nativeContentAd];
-  MPNativeAd *interfaceAd = [[MPNativeAd alloc] initWithAdAdapter:adapter];
-
-  NSMutableArray *imageURLs = [NSMutableArray array];
-
-  if ([interfaceAd.properties[kAdIconImageKey] length]) {
-    if (![MPNativeAdUtils addURLString:interfaceAd.properties[kAdIconImageKey]
-                            toURLArray:imageURLs]) {
-      [self.delegate nativeCustomEvent:self
-              didFailToLoadAdWithError:MPNativeAdNSErrorForInvalidImageURL()];
-    }
-  }
-
-  if ([interfaceAd.properties[kAdMainImageKey] length]) {
-    if (![MPNativeAdUtils addURLString:interfaceAd.properties[kAdMainImageKey]
-                            toURLArray:imageURLs]) {
-      [self.delegate nativeCustomEvent:self
-              didFailToLoadAdWithError:MPNativeAdNSErrorForInvalidImageURL()];
-    }
-  }
-
-  [super precacheImagesWithURLs:imageURLs
-                completionBlock:^(NSArray *errors) {
-                  if (errors) {
-                    [self.delegate nativeCustomEvent:self
-                            didFailToLoadAdWithError:MPNativeAdNSErrorForImageDownloadFailure()];
-                  } else {
-                    [self.delegate nativeCustomEvent:self didLoadAd:interfaceAd];
-                  }
-                }];
-}
-
 #pragma mark - Private Methods
 
-/// Checks the app install ad has required assets or not.
-- (BOOL)isValidAppInstallAd:(GADNativeAppInstallAd *)appInstallAd {
-  return (appInstallAd.headline && appInstallAd.body && appInstallAd.icon &&
-          appInstallAd.images.count && appInstallAd.callToAction);
+/// Checks the unified native ad has required assets or not.
+- (BOOL)isValidUnifiedNativeAd:(GADUnifiedNativeAd *)unifiedNativeAd {
+  return (unifiedNativeAd.headline && unifiedNativeAd.body && unifiedNativeAd.icon &&
+          unifiedNativeAd.images.count && unifiedNativeAd.callToAction);
 }
 
-/// Checks the content ad has required assets or not.
-- (BOOL)isValidContentAd:(GADNativeContentAd *)contentAd {
-  return (contentAd.headline && contentAd.body && contentAd.logo && contentAd.images.count &&
-          contentAd.callToAction);
-}
 @end
+